fix navigation and route in series details
